chore(nutrition): post-merge follow-ups (v0.2.1)
Knocks out 4 of 6 deferred nutrition TODOs plus a tooling fix that
unblocks main's HRT migration.

UI:
- /nutrition now redirects to /nutrition/today. Legacy 937-line page.tsx
  preserved at /nutrition/week as the Week template editor. The legacy
  "Today" subtab inside it stays as a back-door to hydration + day-notes
  editing until the new Today absorbs those.
- Sub-nav points Week at /nutrition/week.

Today page:
- CalorieBalanceCard now shows real burned calories from healthkit_workouts
  via new GET /api/nutrition/today-workouts?date= and useTodayWorkoutCalories
  hook. Replaces the hardcoded 0.

MCP refactor:
- Moved 7 pre-existing nutrition tools (log_nutrition_meal,
  get_active_nutrition_plan, get_nutrition_plan, set_nutrition_day_notes,
  set_nutrition_targets, load_nutrition_plan, update_week_meal) plus the
  DOW_NAMES/parseDayOfWeek helpers from src/lib/mcp-tools.ts into
  src/lib/mcp/nutrition-tools.ts. The whole nutrition MCP surface (15 tools)
  now lives in one file. Main mcp-tools.ts shrinks ~200 lines.

Food search:
- Trigram threshold now explicit per-query (similarity >= 0.22) on both the
  HTTP endpoint and the MCP tool. Per-query beats session-wide so concurrent
  searches don't fight, and 0.22 catches single-character typos in branded
  foods that 0.3 missed.

Migration tooling:
- migrate.ts splitter now recognises -- line comments, /* block comments */
  (with nesting), and 'single-quoted strings'. Previously a semicolon inside
  a comment ("uses the same sync route + state") would split the SQL
  mid-statement. This unblocked main's 020_hrt_labs_meds.sql which has
  applied cleanly to Neon as a result.

TODOs.md updated: 4 marked complete, 2 deferred (materialize view,
real photo+AI integrations), 1 added (absorb hydration into new Today).
