+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rebirth",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "private": true,
   "type": "module",
   "bin": {
