+++ src/app/nutrition/today/page.tsx
@@ -4,6 +4,7 @@ import { Suspense, useEffect, useMemo, useState } from 'react';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { ChevronLeft, ChevronRight, Calendar, Settings } from 'lucide-react';
 import { useNutritionLogsForDate, useDayNote, useNutritionTargets } from '@/lib/useLocalDB-nutrition';
+import { useTodayWorkoutCalories } from '@/lib/useTodayWorkouts';
 import { deleteMeal } from '@/lib/mutations-nutrition';
 import {
   todayLocal,
@@ -62,6 +63,7 @@ function NutritionTodayContent() {
   const logs = useMemo(() => rawLogs ?? [], [rawLogs]);
   const dayNote = useDayNote(date);
   const targets = useNutritionTargets();
+  const workoutKcal = useTodayWorkoutCalories(date);
 
   const bands = (targets?.bands ?? DEFAULT_BANDS) as MacroBands;
 
@@ -143,7 +145,7 @@ function NutritionTodayContent() {
       <CalorieBalanceCard
         consumed={totals.cal}
         goal={targets?.calories ?? null}
-        workouts={0}
+        workouts={workoutKcal}
         band={bands.cal ?? null}
       />
