+++ src/lib/mcp/nutrition-tools.ts
@@ -24,11 +24,25 @@
  *     day where the user is expected to actively approve.
  */
 
-import { query, queryOne } from '@/db/db';
+import { query, queryOne, transaction } from '@/db/db';
 import { searchOpenFoodFacts, searchUsdaFdc } from '@/lib/food-search-remote';
 import { computeDayAdherence, computeStreak, DEFAULT_BANDS } from '@/lib/adherence';
 import type { MacroBands } from '@/db/local';
 import type { MCPTool } from '@/lib/mcp-tools';
+
+// ─── Shared day-of-week helpers (used by load_nutrition_plan + get_nutrition_plan) ──
+
+const DOW_NAMES = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'];
+
+const DAY_NAME_MAP: Record<string, number> = {
+  mon: 0, tue: 1, wed: 2, thu: 3, fri: 4, sat: 5, sun: 6,
+};
+
+function parseDayOfWeek(day: string | number): number {
+  if (typeof day === 'number') return Math.max(0, Math.min(6, day - 1));
+  const key = String(day).toLowerCase().slice(0, 3);
+  return DAY_NAME_MAP[key] ?? 0;
+}
 
 // ─── Result helpers (mirror mcp-tools.ts) ────────────────────────────────────
 
@@ -310,7 +324,7 @@ async function searchNutritionFoods(args: Record<string, unknown>) {
          FROM nutrition_food_canonical
          WHERE canonical_name LIKE $1 || '%' ESCAPE '\\'
             OR canonical_name LIKE '%' || $1 || '%' ESCAPE '\\'
-            OR canonical_name % $2
+            OR similarity(canonical_name, $2) >= 0.22
          ORDER BY (canonical_name LIKE $1 || '%' ESCAPE '\\') DESC,
                   times_logged DESC, last_logged_at DESC
          LIMIT $3`,
@@ -474,6 +488,266 @@ async function getNutritionSummary(args: Record<string, unknown>) {
 }
 
 // ─── 8. get_nutrition_rules ──────────────────────────────────────────────────
+// (Pre-existing tools — moved out of mcp-tools.ts so the entire nutrition
+// MCP surface lives in one file.)
+
+// get_active_nutrition_plan
+async function getActiveNutritionPlan() {
+  const meals = await query(`
+    SELECT uuid, day_of_week, meal_slot, meal_name,
+           protein_g, carbs_g, fat_g, calories, quality_rating, sort_order
+    FROM nutrition_week_meals
+    ORDER BY day_of_week, sort_order
+  `);
+  return toolResult(meals);
+}
+
+// get_nutrition_plan
+async function getNutritionPlan(args: Record<string, unknown>) {
+  const includeCompliance = args.include_compliance === true;
+
+  const meals = await query<{
+    uuid: string; day_of_week: number; meal_slot: string; meal_name: string;
+    protein_g: number | null; carbs_g: number | null; fat_g: number | null;
+    calories: number | null; quality_rating: number | null; sort_order: number;
+  }>(`
+    SELECT uuid, day_of_week, meal_slot, meal_name,
+           protein_g, carbs_g, fat_g, calories, quality_rating, sort_order
+    FROM nutrition_week_meals
+    ORDER BY day_of_week, sort_order
+  `);
+
+  const byDay = new Map<number, typeof meals>();
+  for (const m of meals) {
+    if (!byDay.has(m.day_of_week)) byDay.set(m.day_of_week, []);
+    byDay.get(m.day_of_week)!.push(m);
+  }
+
+  const week_plan = Array.from(byDay.entries())
+    .sort(([a], [b]) => a - b)
+    .map(([dow, dayMeals]) => ({
+      day: DOW_NAMES[dow] ?? String(dow),
+      meals: dayMeals.map(m => ({
+        uuid: m.uuid,
+        slot: m.meal_slot,
+        description: m.meal_name,
+        calories: m.calories,
+        protein_g: m.protein_g,
+        carbs_g: m.carbs_g,
+        fat_g: m.fat_g,
+        quality_rating: m.quality_rating,
+      })),
+    }));
+
+  const targets = await queryOne<{
+    calories: number | null; protein_g: number | null;
+    carbs_g: number | null; fat_g: number | null;
+  }>(`SELECT calories, protein_g, carbs_g, fat_g FROM nutrition_targets WHERE id = 1`);
+
+  const result: Record<string, unknown> = { week_plan, targets: targets ?? null };
+
+  if (includeCompliance) {
+    const compliance = await queryOne<{
+      avg_calories: string | null; avg_protein: string | null; logged_days: string;
+    }>(`
+      SELECT
+        ROUND(AVG(daily_cal)::numeric, 1) AS avg_calories,
+        ROUND(AVG(daily_prot)::numeric, 1) AS avg_protein,
+        COUNT(*)::int AS logged_days
+      FROM (
+        SELECT
+          DATE(logged_at) AS d,
+          SUM(calories) AS daily_cal,
+          SUM(protein_g) AS daily_prot
+        FROM nutrition_logs
+        WHERE logged_at >= NOW() - INTERVAL '7 days'
+        GROUP BY DATE(logged_at)
+      ) sub
+    `);
+    result.compliance_7d = {
+      avg_calories: compliance?.avg_calories != null ? Number(compliance.avg_calories) : null,
+      avg_protein: compliance?.avg_protein != null ? Number(compliance.avg_protein) : null,
+      logged_days: Number(compliance?.logged_days ?? 0),
+    };
+  }
+
+  return toolResult(result);
+}
+
+// load_nutrition_plan
+async function loadNutritionPlan(args: Record<string, unknown>) {
+  const { days, targets } = args as {
+    days: Array<{
+      day: string | number;
+      meals: Array<{
+        slot: string; description: string; calories?: number;
+        protein_g?: number; carbs_g?: number; fat_g?: number;
+        quality_rating?: number;
+      }>;
+    }>;
+    targets?: { calories?: number; protein_g?: number; carbs_g?: number; fat_g?: number };
+  };
+
+  const statements: Array<{ text: string; params?: unknown[] }> = [
+    { text: 'DELETE FROM nutrition_week_meals' },
+  ];
+
+  let mealCount = 0;
+  for (const d of days) {
+    const dow = parseDayOfWeek(d.day as string | number);
+    for (let i = 0; i < d.meals.length; i++) {
+      const m = d.meals[i];
+      statements.push({
+        text: `INSERT INTO nutrition_week_meals
+           (uuid, day_of_week, meal_slot, meal_name,
+            protein_g, carbs_g, fat_g, calories, quality_rating, sort_order)
+         VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)`,
+        params: [
+          crypto.randomUUID(), dow, m.slot, m.description,
+          m.protein_g ?? null, m.carbs_g ?? null, m.fat_g ?? null,
+          m.calories ?? null, m.quality_rating ?? null, i,
+        ],
+      });
+      mealCount++;
+    }
+  }
+
+  if (targets) {
+    statements.push({
+      text: `INSERT INTO nutrition_targets (id, calories, protein_g, carbs_g, fat_g, updated_at)
+       VALUES (1, $1, $2, $3, $4, NOW())
+       ON CONFLICT (id) DO UPDATE SET
+         calories = EXCLUDED.calories,
+         protein_g = EXCLUDED.protein_g,
+         carbs_g = EXCLUDED.carbs_g,
+         fat_g = EXCLUDED.fat_g,
+         updated_at = NOW()`,
+      params: [targets.calories ?? null, targets.protein_g ?? null, targets.carbs_g ?? null, targets.fat_g ?? null],
+    });
+  }
+
+  await transaction(statements);
+
+  return toolResult({
+    success: true,
+    meals_created: mealCount,
+    weeks_loaded: days.length,
+    targets_set: targets ? true : false,
+  });
+}
+
+const LOG_MEAL_TYPES = ['breakfast', 'lunch', 'dinner', 'snack', 'other'] as const;
+const LOG_MEAL_STATUSES = ['planned', 'deviation', 'added'] as const;
+
+// log_nutrition_meal
+async function logNutritionMeal(args: Record<string, unknown>) {
+  const {
+    meal_type = null, meal_name = null, calories, protein_g, carbs_g, fat_g,
+    notes = null, template_meal_id = null, status = null, logged_at,
+  } = args as {
+    meal_type?: string | null; meal_name?: string | null;
+    calories?: number; protein_g?: number; carbs_g?: number; fat_g?: number;
+    notes?: string | null; template_meal_id?: string | null;
+    status?: string | null; logged_at?: string;
+  };
+
+  if (meal_type != null && !LOG_MEAL_TYPES.includes(meal_type as typeof LOG_MEAL_TYPES[number])) {
+    return toolError('INVALID_INPUT', `meal_type must be one of ${LOG_MEAL_TYPES.join(', ')}`);
+  }
+  if (status != null && !LOG_MEAL_STATUSES.includes(status as typeof LOG_MEAL_STATUSES[number])) {
+    return toolError('INVALID_INPUT', `status must be one of ${LOG_MEAL_STATUSES.join(', ')}`);
+  }
+
+  const row = await queryOne(
+    `INSERT INTO nutrition_logs
+       (uuid, logged_at, meal_type, meal_name, calories, protein_g, carbs_g, fat_g,
+        notes, template_meal_id, status)
+     VALUES ($1, COALESCE($2::timestamp, NOW()), $3, $4, $5, $6, $7, $8, $9, $10, $11)
+     RETURNING *`,
+    [
+      crypto.randomUUID(), logged_at ?? null, meal_type, meal_name,
+      calories ?? null, protein_g ?? null, carbs_g ?? null, fat_g ?? null,
+      notes, template_meal_id, status,
+    ],
+  );
+  return toolResult(row);
+}
+
+// set_nutrition_day_notes
+async function setNutritionDayNotes(args: Record<string, unknown>) {
+  const { date, hydration_ml = null, notes = null } = args as {
+    date?: string; hydration_ml?: number | null; notes?: string | null;
+  };
+  if (typeof date !== 'string' || !DATE_RE.test(date)) {
+    return toolError('INVALID_INPUT', 'date is required in YYYY-MM-DD format');
+  }
+
+  const row = await queryOne(
+    `INSERT INTO nutrition_day_notes (uuid, date, hydration_ml, notes, updated_at)
+     VALUES ($1, $2, $3, $4, NOW())
+     ON CONFLICT (date) DO UPDATE SET
+       hydration_ml = COALESCE(EXCLUDED.hydration_ml, nutrition_day_notes.hydration_ml),
+       notes = COALESCE(EXCLUDED.notes, nutrition_day_notes.notes),
+       updated_at = NOW()
+     RETURNING *`,
+    [crypto.randomUUID(), date, hydration_ml, notes],
+  );
+  return toolResult(row);
+}
+
+// set_nutrition_targets
+async function setNutritionTargetsTool(args: Record<string, unknown>) {
+  const { calories = null, protein_g = null, carbs_g = null, fat_g = null } = args as {
+    calories?: number | null; protein_g?: number | null;
+    carbs_g?: number | null; fat_g?: number | null;
+  };
+
+  const row = await queryOne(
+    `INSERT INTO nutrition_targets (id, calories, protein_g, carbs_g, fat_g, updated_at)
+     VALUES (1, $1, $2, $3, $4, NOW())
+     ON CONFLICT (id) DO UPDATE SET
+       calories = EXCLUDED.calories,
+       protein_g = EXCLUDED.protein_g,
+       carbs_g = EXCLUDED.carbs_g,
+       fat_g = EXCLUDED.fat_g,
+       updated_at = NOW()
+     RETURNING *`,
+    [calories, protein_g, carbs_g, fat_g],
+  );
+  return toolResult(row);
+}
+
+// update_week_meal
+async function updateWeekMeal(args: Record<string, unknown>) {
+  const { uuid } = args;
+  if (typeof uuid !== 'string') return toolError('INVALID_INPUT', 'uuid is required');
+
+  const fields: string[] = [];
+  const params: unknown[] = [];
+  let p = 0;
+  const pushField = (col: string, val: unknown) => {
+    fields.push(`${col} = $${++p}`);
+    params.push(val);
+  };
+
+  if (typeof args.meal_slot === 'string') pushField('meal_slot', args.meal_slot);
+  if (typeof args.meal_name === 'string') pushField('meal_name', args.meal_name);
+  if (args.calories !== undefined) pushField('calories', args.calories ?? null);
+  if (args.protein_g !== undefined) pushField('protein_g', args.protein_g ?? null);
+  if (args.carbs_g !== undefined) pushField('carbs_g', args.carbs_g ?? null);
+  if (args.fat_g !== undefined) pushField('fat_g', args.fat_g ?? null);
+  if (args.quality_rating !== undefined) pushField('quality_rating', args.quality_rating ?? null);
+  if (typeof args.sort_order === 'number') pushField('sort_order', args.sort_order);
+
+  if (fields.length === 0) return toolError('INVALID_INPUT', 'No fields to update');
+
+  params.push(uuid);
+  const row = await queryOne(
+    `UPDATE nutrition_week_meals SET ${fields.join(', ')} WHERE uuid = $${++p} RETURNING *`,
+    params,
+  );
+  return row ? toolResult(row) : toolError('NOT_FOUND', 'Meal not found');
+}
 
 async function getNutritionRules() {
   return toolResult({
@@ -626,5 +900,141 @@ export const nutritionTools: MCPTool[] = [
       'Returns the implicit rules that govern the nutrition surface: auto-approval semantics, date/timezone conventions, default adherence bands, and recommended tool workflows. Call this once when first interacting with nutrition tools.',
     inputSchema: { type: 'object', properties: {} },
     execute: getNutritionRules,
+  },
+
+  // ─── Pre-existing tools (relocated from mcp-tools.ts) ──────────────────────
+
+  {
+    name: 'get_active_nutrition_plan',
+    description: 'Returns the current standard-week meal plan template.',
+    inputSchema: { type: 'object', properties: {} },
+    execute: async () => getActiveNutritionPlan(),
+  },
+  {
+    name: 'get_nutrition_plan',
+    description: 'Returns the current weekly meal plan grouped by day, with optional 7-day compliance summary from nutrition logs.',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        include_compliance: { type: 'boolean', description: 'If true, include avg calories/protein from the last 7 logged days' },
+      },
+    },
+    execute: getNutritionPlan,
+  },
+  {
+    name: 'load_nutrition_plan',
+    description: 'Replaces the entire standard-week meal template with a new plan. Accepts days with nested meals and optional daily macro targets.',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        days: {
+          type: 'array',
+          description: 'Array of days with their meals',
+          items: {
+            type: 'object',
+            properties: {
+              day: { type: ['string', 'number'], description: 'Day name (Mon/Tue/Wed/Thu/Fri/Sat/Sun) or number (1=Mon … 7=Sun)' },
+              meals: {
+                type: 'array',
+                items: {
+                  type: 'object',
+                  properties: {
+                    slot: { type: 'string', description: 'breakfast | lunch | dinner | snack' },
+                    description: { type: 'string', description: 'Meal name/description' },
+                    calories: { type: 'number' },
+                    protein_g: { type: 'number' },
+                    carbs_g: { type: 'number' },
+                    fat_g: { type: 'number' },
+                    quality_rating: { type: 'number', description: '1–5 quality rating' },
+                  },
+                  required: ['slot', 'description'],
+                },
+              },
+            },
+            required: ['day', 'meals'],
+          },
+        },
+        targets: {
+          type: 'object',
+          description: 'Daily macro targets (upserts the singleton nutrition_targets row).',
+          properties: {
+            calories: { type: 'number' },
+            protein_g: { type: 'number' },
+            carbs_g: { type: 'number' },
+            fat_g: { type: 'number' },
+          },
+        },
+      },
+      required: ['days'],
+    },
+    execute: loadNutritionPlan,
+  },
+  {
+    name: 'update_week_meal',
+    description: "Partially updates a single meal in the standard-week template by uuid. Only provided fields are changed.",
+    inputSchema: {
+      type: 'object',
+      properties: {
+        uuid: { type: 'string', description: 'UUID of the nutrition_week_meals row' },
+        meal_slot: { type: 'string', description: 'breakfast | lunch | dinner | snack' },
+        meal_name: { type: 'string' },
+        calories: { type: 'number' },
+        protein_g: { type: 'number' },
+        carbs_g: { type: 'number' },
+        fat_g: { type: 'number' },
+        quality_rating: { type: 'number', description: '1–5' },
+        sort_order: { type: 'number' },
+      },
+      required: ['uuid'],
+    },
+    execute: updateWeekMeal,
+  },
+  {
+    name: 'log_nutrition_meal',
+    description: 'Logs an actual eaten meal to nutrition_logs (compliance + deviation tracking). meal_type one of breakfast|lunch|dinner|snack|other; status one of planned|deviation|added.',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        meal_type: { type: 'string', description: 'breakfast | lunch | dinner | snack | other' },
+        meal_name: { type: 'string' },
+        calories: { type: 'number' },
+        protein_g: { type: 'number' },
+        carbs_g: { type: 'number' },
+        fat_g: { type: 'number' },
+        notes: { type: 'string' },
+        template_meal_id: { type: 'string', description: 'UUID of the nutrition_week_meals row this log came from' },
+        status: { type: 'string', description: 'planned | deviation | added' },
+        logged_at: { type: 'string', description: 'ISO timestamp (defaults to now)' },
+      },
+    },
+    execute: logNutritionMeal,
+  },
+  {
+    name: 'set_nutrition_day_notes',
+    description: 'Upserts hydration and free-text notes for a specific calendar day (YYYY-MM-DD). Omitted fields preserve existing values.',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        date: { type: 'string', description: 'YYYY-MM-DD' },
+        hydration_ml: { type: 'number' },
+        notes: { type: 'string' },
+      },
+      required: ['date'],
+    },
+    execute: setNutritionDayNotes,
+  },
+  {
+    name: 'set_nutrition_targets',
+    description: 'Replaces the singleton daily-macro-targets row. Omitted fields are set to null — always pass the full set (calories, protein_g, carbs_g, fat_g) if you want all four tracked.',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        calories: { type: 'number' },
+        protein_g: { type: 'number' },
+        carbs_g: { type: 'number' },
+        fat_g: { type: 'number' },
+      },
+    },
+    execute: setNutritionTargetsTool,
   },
 ];
